Add new `listParallel` method; `tailMask` compares < d entries
The `tailMask` property allows equality comparisons to be done only
across the first `tailMask` (< d) entries in a vector; this also affects
how vectors are normalized through the `dirIndep` setting.

The objective of this change is to allow testing for mere parallel-ness
of vectors when they are interpreted as specifying a hyperplane in d-1
dimensions. This is accomplished via the new `listParallel` method.

`listParallel` takes a "query" vector as input, and returns two lists of
indexes: one contains those "hyperplanes" that are exactly equal to the
query (up to normalization); the other contains a list of those
hyperplanes that are equal to the query *only along the first
`tailMask=` dimensions* up to scaling. When the `tailMask=` optional
argument is set to 1, this amounts to testing hyperplanes for being
merely parallel to the query.

In this commit, computation of the `self.scale` property was also made
more efficient.